+++ src/lib.rs
@@ -25,7 +25,10 @@ mod cmp;
 mod const_for;
 mod div;
 mod from;
+mod log;
 mod mul;
+mod pow;
+mod special;
 mod string;
 mod support;
 mod uint_dyn;
@@ -77,8 +80,27 @@ pub mod nightly {
 }
 
 /// The ring of numbers modulo $2^{\mathtt{BITS}}$.
-// TODO: Get rid of the `LIMBS` argument when  `generic_const_exprs` stabilizes.
-// Blocked by Rust [#76560](https://github.com/rust-lang/rust/issues/76560).
+///
+/// [`Uint`] implements nearly all traits and methods from the `std` unsigned
+/// integer types, including most nightly only ones.
+///
+/// # Notable differences from `std` uint types.
+///
+/// * The operators `+`, `-`, `*`, etc. using wrapping math by default. The std
+///   operators panic on overflow in debug, and are undefined in release, see
+///   [reference][std-overflow].
+/// * The [`Uint::checked_shl`], [`Uint::overflowing_shl`], etc return overflow
+///   when non-zero bits are shifted out. In std they return overflow when the
+///   shift amount is greater than the bit size.
+/// * Some methods like [`u64::div_euclid`] and [`u64::rem_euclid`] are left out
+///   because they are meaningless or redundant for unsigned integers. Std has
+///   them for compatibility with their signed integers.
+/// * Many functions that are `const` in std are not in [`Uint`].
+/// * [`Uint::to_le_bytes`] and [`Uint::to_be_bytes`] require the output size to
+///   be provided as a const-generic argument. They will runtime panic if the
+///   provided size is incorrect.
+///
+/// [std-overflow]: https://doc.rust-lang.org/reference/expressions/operator-expr.html#overflow
 #[derive(Clone, Copy, Eq, PartialEq, Hash)]
 pub struct Uint<const BITS: usize, const LIMBS: usize> {
     limbs: [u64; LIMBS],
@@ -241,5 +263,7 @@ pub mod bench {
         add::bench::group(criterion);
         mul::bench::group(criterion);
         div::bench::group(criterion);
+        pow::bench::group(criterion);
+        log::bench::group(criterion);
     }
 }