+++ src/support/mod.rs
@@ -10,6 +10,7 @@ mod serde;
 
 // TODO:
 // * https://docs.rs/tracing/latest/tracing/trait.Value.html
+// * https://crates.io/crates/num-traits
 // * https://crates.io/crates/der
 // * https://crates.io/crates/zeroize
 // * https://crates.io/crates/ff